Move remaining makeboxes format rules to their corresponding symbols in mathics.builtin.layout (#1638)
* Move MakeBoxes rules associated to symbols in Layout to the
corresponding symbols.
* Fix PrecedenceForm in OutputForm.
* Fix Optional makeboxes rules.
* Add some extra tests in format_tests.yaml.

Intermediate stage before converting 'MakeBoxes' rules into
'FormatValues'

---------
